updater/pyproject.toml
[tool.poetry]
name = "updater"
version = "0.1.0"
description = "Updater periodically polls Notion databases and save data to local database"
authors = ["Alexey Serikov <serikov.alexey42@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
SQLAlchemy = "^2.0.28"
asyncpg = "^0.29.0"
venusian = "^3.1.0"
isort = "^5.13.2"
black = "^24.2.0"
flake8 = "^7.0.0"
notion-client = "^2.2.1"
pydantic-settings = "^2.2.1"
traceback-with-variables = "^2.0.4"
alembic = "^1.13.1"
aiogram = "^3.6.0"
aio-pika = "^9.4.1"
deepdiff = "^7.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
